Minor optimizations in Plate Solving page
