+++ frontend/src/PlateSolving/PlateSolvingContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
-import { PlateSolvingPage, PlateSolvingSectionMenu } from './PlateSolvingPage';
+import { PlateSolving } from './PlateSolving';
 import Actions from '../actions';
-import { plateSolvingContainerSelector, plateSolvingSectionMenuSelector } from './selectors';
+import { plateSolvingContainerSelector } from './selectors';
 
 const mapDispatchToProps = dispatch => ({
   setOption: (option, value) => dispatch(Actions.PlateSolving.setOption(option, value)),
@@ -11,11 +11,6 @@ const mapDispatchToProps = dispatch => ({
 export const PlateSolvingContainer = connect(
   plateSolvingContainerSelector,
   mapDispatchToProps
-)(PlateSolvingPage)
-
-export const PlateSolvingSectionMenuContainer = connect(
-  plateSolvingSectionMenuSelector,
-  null, 
-)(PlateSolvingSectionMenu)
+)(PlateSolving)
